Refactor, part 1
